add : 搜索功能添加

+++ ZJCityPickerViewController/ZJCityPickerViewController/ZJCityPickerViewController/ZJCityPickerAppearance.m
@@ -54,7 +54,7 @@
     return [self calculateHeaderHeightWithItemsCount:groupModel.cityArray.count];
 }
 
-- (CGFloat)calculateHeaderHeightWithItemsCount:(CGFloat)itemsCount {
+- (CGFloat)calculateHeaderHeightWithItemsCount:(NSInteger )itemsCount {
     CGFloat sectionHeight = _sectionInset.top + _sectionInset.bottom + _headerHeight;
     CGFloat columns = (itemsCount - 1) / 3 + 1;
     CGFloat itemsHeight = columns * _itemHeight + (columns > 0 ? (columns - 1) * _itemMinLineSpace : 0);
